separating out the Argument components
